# ARROW-12343: [Rust] Support auto-vectorization for min/max