--- a/rust/arrow/src/compute/kernels/aggregate.rs
+++ b/rust/arrow/src/compute/kernels/aggregate.rs
@@ -107,7 +107,6 @@ pub fn min_string<T: StringOffsetSizeTrait>(
 
 #[multiversion]
 #[clone(target = "x86_64+avx")]
-#[clone(target = "x86_64+avx+avx2")]
 fn min_max_helper<T, F>(array: &PrimitiveArray<T>, cmp: F) -> Option<T::Native>
 where
     T: ArrowNumericType,