@Dandandan 2021-04-12
@jorgecarleitao 

It definitely not the same as when compiling with the `simd` feature, that is about 4 as fast as this version (and ~5x faster than the master branch). This is with `--features simd`:

```
min 512                 time:   [232.68 ns 233.24 ns 233.83 ns]                    
                        change: [-75.848% -75.631% -75.432%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) low mild
  1 (1.00%) high mild
  5 (5.00%) high severe

max 512                 time:   [229.37 ns 229.71 ns 230.13 ns]                    
                        change: [-75.484% -75.392% -75.303%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  11 (11.00%) high mild
  7 (7.00%) high severe
```