@alamb 2021-04-14
> holy molly, 6 lines of code. Brilliant.

I 100% agree with this sentiment.