@alamb 2021-04-15
@nevi-me  / @jorgecarleitao  any concerns about merging this PR? Only the new dependency is of potential concern to me, but it seems to me to be a pretty reasonable tradeoff

We could probably also just inline the part of the `multiversion` crate we needed (I bet it has more features than we are using) but it isn't clear to me avoiding the dependency in this case is worth it