@alamb 2021-04-17
It sounds to me like there is consensus on this approach. I'll plan to merge it in as soon as the 4.0 release is out the door (ETA early next week) so as not to potentially introduce instability into the release process at this stage