@Dandandan 2021-04-17
> The multiversion crate seems really handy! Are the tradeoffs just the expected bump in binary size to support the multiple implementations and presumably a branch when entering a multi-target code path? Is there any facility to remove the runtime checks, e.g. if I know I'm producing builds for say aarch64?

AFAIK the architecture is static/compile time, so it will not be included in the checks. If the architecture is not the target architecture, it shouldn't include the branches for the features and in the generated code.

> My only concern would be on binary size if we really go crazy using it but we can monitor that over time.

Agreed. Good to monitor this and try to keep size down :+1: